// video-games/src/models/videoGames.interface.ts
export interface Course {
  id: number;
  groups: any[];
  thumbnail: string;
  cover?: string;
  final_project_cover: string;
  jumbotron_cover: string;
  jumbotron_cover_mobile: string;
  name: string;
  slug: string;
  youtube_video_id: string;
  call_out: string;
  description: string;
  title: string;
  course_detail: string;
  order_in_list: number;
  order_in_learning_path: number;
  number_of_sessions: number;
  age_range: string;
  business_type: string;
  frequency: number;
  duration: string;
  skills_to_develop: string[];
  learnings: Learning[];
  final_project_description: string;
  project_summary: string;
  technical_requirements: string[];
  lessons: string[];
  modules: any[];
  main_color: string;
  curriculum: string;
  instructive: string;
  has_discount: boolean;
  show_as_free_course: boolean;
  course_base_mdl_id: number;
  category_mdl_id: number;
  has_test_month: boolean;
  amount_usd: string;
  amount_pen: string;
  amount_mxn: string;
  amount_col: string;
  discount_percentage: string;
  amount_after_discount_usd: string;
  amount_after_discount_pen: string;
  amount_after_discount_mxn: string;
  amount_after_discount_col: string;
  test_month_amount: string;
  test_month_amount_usd: string;
  test_month_amount_mxn: string;
  test_month_amount_col: string;
  test_month_duration: number;
  test_month_classes: number;
  benefits: any[];
  has_seo_noindex: boolean;
  course_wiki_url?: any;
  available_to_purchase: boolean;
  web_star: boolean;
  bundle?: any;
  campaign?: any;
  age_range_v2: number[];
}

export interface Learning {
  title: string;
  description: string;
}